requirements.txt
flask
kafka-python
pyspark
elasticsearch
pymongo
pdfplumber